Fix uncertain edit ergonomics measurement
